Avoid column progression subpixel inaccuracies.
We can use LayoutUnitDiffuser so that everything sums up nicely.

Update a WPT test that relied on the inaccuracies now fixed. With the
code changes in this CL, we would render identically to what Firefox
does, but that wasn't the expected result, even though it is better.
Rewrite the test to not rely on subpixel behavior.

Also update one unit test, which was using subpixels for no reason (and
relying on the results being inaccurate).

Rewrite css-multicol/multicol-width-ch-001.xht , to make it fit within
the viewport (it was much wider than 800px), not use XHTML, and use grid
in the ref, since that's hopefully more likely to match. Adding five
13ch width elements with 1ch between them is not guaranteed to be the
same as 69ch. Instead, use grid, and distribute 1ch to the "gaps" and
1fr for the five "columns". And use Ahem.

Bug: 467869214
Change-Id: I946d2cd39e78b987af402c43f29fff191634786b